lib-manager: Support for two stage library loading
The hardware programming flow dictates that the DSP side should set the
GEN bit first followed by the host setting the EN bit.
We cannot assume that the GEN bit is already enabled (and it would in fact
will leave the DMA on the DSP side wrongly configured at best).

To support the currently used flow (single step) and add support for the
correct programming flow, the proposed solution is to use the lib_id as
indication of the method the host software is going to use.

Two stage flow:
The first LOAD_LIBRARY with lib_id = 0 and valid dma_id
 - just allocate and set the GEN bit
Second LOAD_LIBRARY with valid lib_id and dma_id
 - proceed to the loading of the library via host DMA to local buffer

Single stage flow (currently supported)
LOAD_LIBRARY with valid lib_id and dma_id
 - allocate, set GEN bit and proceed to loading the library

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>